refactor(dispatch): extract envelope_govern_support.py leaf module (PR-3 monolith split)
Moves seven GOVERN-seam support functions unchanged from dispatch_envelope.py
to a new envelope_govern_support.py leaf module: _receipt_exists_for_dispatch,
_resolve_fix_forward_diff, _resolve_phantom_diff, _unknown_verification,
_verification_from_report, _archive_dispatch_events, _clear_dispatch_events.

The facade imports them via bare name (`from envelope_govern_support import X`)
so every `patch("dispatch_envelope.X")`-shaped test coupling that targets a
caller still in the facade (_govern, unmoved until PR-4) keeps binding.

_receipt_exists_for_dispatch's logger name changes from "dispatch_envelope" to
"envelope_govern_support" with the move; the two caplog captures on the old
name in tests/test_dispatch_envelope.py (lines 557, 591) and the two
equivalent captures in test_dispatch_envelope_characterization.py move with
it in this same commit. The four _archive_dispatch_events/_clear_dispatch_events
patch sites (test_dispatch_envelope.py:236-237, test_role_applied_provider_lane.py:211-212)
are intentionally left targeting dispatch_envelope — their caller (_govern)
stays in the facade until PR-4.

tests/data/dispatch_envelope_census.json regenerated in this commit via
`python3 tests/dispatch_envelope_census_scanner.py` to reflect the new
envelope_govern_support family member and the relocated logger-name/Layer-4
couplings.

Dispatch-ID: 20260803-223537-pr3-envelope-govern-support
