+++ tests/test_dispatch_envelope.py
@@ -554,7 +554,7 @@ class TestEnvelopeIdempotentDedup:
             return _real_open(path, *args, **kwargs)
 
         import logging
-        with caplog.at_level(logging.WARNING, logger="dispatch_envelope"):
+        with caplog.at_level(logging.WARNING, logger="envelope_govern_support"):
             with patch("provider_spawns.claude_spawn.spawn_claude", return_value=claude_result), \
                  patch("governance_emit.emit_unified_report", mock_report), \
                  patch("governance_emit.emit_dispatch_receipt", mock_receipt), \
@@ -588,7 +588,7 @@ class TestEnvelopeIdempotentDedup:
             return _real_open(path, *args, **kwargs)
 
         import logging
-        with caplog.at_level(logging.WARNING, logger="dispatch_envelope"):
+        with caplog.at_level(logging.WARNING, logger="envelope_govern_support"):
             with patch("builtins.open", side_effect=_raise_oserror):
                 result = dispatch_envelope._receipt_exists_for_dispatch(
                     receipt_path, "some-id"
